+++ src/main/resources/intentionDescriptions/LuaInvertIfIntention/after.template.lua
@@ -1,0 +1,5 @@
+if x ~= 1 then
+    bar()
+else
+    foo()
+end